pyproject.toml
[tool.poetry]
name = "cnb-cluer"
version = "0.1.0"
description = ""
authors = ["Nalu Zou"]
readme = "README.md"
packages = [{include = "cnb_cluer"}]

[tool.poetry.dependencies]
python = "^3.10"
gensim = "^4.2.0"
tqdm = "^4.64.1"
openai = "^0.26.2"
PyYAML = "^6.0"

[tool.poetry.scripts]
generate-clues  = "cnb_cluer.generate_clues:main"
get-guesses  = "cnb_cluer.get_guesses:main"
evaluate-combos = "cnb_cluer.evaluate_combos:main"

test-gensim     = "tools.test_gensim:main"
label-types     = "tools.label_types:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
